import: if gfile.trash() returns false, also fall back on gfile.delete()

+++ src/paperwork/frontend/mainwindow/__init__.py
@@ -1655,13 +1655,15 @@ class ActionImport(SimpleAction):
         for file_uri in file_uris:
             logger.info("Moving {} to trash ...".format(file_uri))
             gfile = Gio.File.new_for_uri(file_uri)
+            trashed = False
             try:
-                gfile.trash()
+                trashed = gfile.trash()
             except Exception as exc:
                 logger.warning(
                     "Failed to move file %s to trash. Will delete it",
                     file_uri, exc_info=exc
                 )
+            if not trashed:
                 gfile.delete()
         notification = Notify.Notification.new(
             _("Imported file(s) deleted"),
